remove components/renku_data_services/secrets/low_level_db.py and fix import cycle
