+++ components/renku_data_services/secrets/core.py
@@ -2,6 +2,7 @@
 
 import asyncio
 from base64 import b64encode
+from typing import TYPE_CHECKING
 
 from cryptography.hazmat.primitives.asymmetric import rsa
 from kubernetes import client as k8s_client
@@ -12,7 +13,6 @@ from ulid import ULID
 from renku_data_services import base_models, errors
 from renku_data_services.base_models.core import InternalServiceAdmin
 from renku_data_services.k8s.client_interfaces import K8sCoreClientInterface
-from renku_data_services.secrets.low_level_db import LowLevelUserSecretsRepo
 from renku_data_services.secrets.models import OwnerReference, Secret
 from renku_data_services.users.db import UserRepo
 from renku_data_services.utils.cryptography import (
@@ -23,6 +23,9 @@ from renku_data_services.utils.cryptography import (
     generate_random_encryption_key,
 )
 
+if TYPE_CHECKING:
+    from renku_data_services.secrets.db import LowLevelUserSecretsRepo
+
 
 async def create_k8s_secret(
     user: base_models.APIUser,
@@ -30,7 +33,7 @@ async def create_k8s_secret(
     namespace: str,
     secret_ids: list[ULID],
     owner_references: list[OwnerReference],
-    secrets_repo: LowLevelUserSecretsRepo,
+    secrets_repo: "LowLevelUserSecretsRepo",
     secret_service_private_key: rsa.RSAPrivateKey,
     previous_secret_service_private_key: rsa.RSAPrivateKey | None,
     core_client: K8sCoreClientInterface,
@@ -112,7 +115,7 @@ async def rotate_encryption_keys(
     requested_by: InternalServiceAdmin,
     new_key: rsa.RSAPrivateKey,
     old_key: rsa.RSAPrivateKey,
-    secrets_repo: LowLevelUserSecretsRepo,
+    secrets_repo: "LowLevelUserSecretsRepo",
     batch_size: int = 100,
 ) -> None:
     """Rotate all secrets to a new private key.
